% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{getMedDraTerms}
\alias{getMedDraTerms}
\title{Get MedDRA Terms}
\usage{
getMedDraTerms(sourceCodeDf, baseUrl)
}
\arguments{
\item{sourceCodeDf}{A data frame consisting of SOURCE_CODE, SOURCE_VOCABULARY_ID}

\item{baseUrl}{The WebAPI endpoint}
}
\description{
Get MedDRA Terms
}
